moved get_entity_class to helpers and improved typing
